scripts/requirements.txt
azure-identity==1.13.0b3
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
